pyproject.toml
[project]
name = "kairos"
version = "0.6.2"
description = ""
authors = [
    {name = "Jasper Cantwell",email = "jasper66018@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "pydantic (>=2.11.7,<3.0.0)",
    "fastapi (>=0.115.13,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pydantic-settings (>=2.10.0,<3.0.0)",
    "pymongo (>=4.13.2,<5.0.0)",
    "pydantic-extra-types (>=2.10.5,<3.0.0)",
    "mangum (>=0.19.0,<0.20.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "resend (>=2.14.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
src_paths = ["app"]
[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"

